fix [attachment]: fix delete plugin

+++ backend/internal/service/upload/interface.go
@@ -10,11 +10,11 @@ import (
 type Engine interface {
 	// Upload 仅负责存储文件，返回存储结果，不操作数据库
 	Upload(ctx context.Context, req *UploadRequest) (*UploadResult, error)
-	DeleteFile(ctx context.Context, storedPath string) error   // 新增
+	DeleteFile(ctx context.Context, storedPath string) error 
 }
 
 type UploadRequest struct {
-	UserID    int64
+	UserID    uint
 	PluginID  string
 	File      *multipart.FileHeader
 	FileType  do.FileType
